webhooks/project_requirements.txt
python-dotenv>=0.15.0
sqlalchemy>=1.4.32
pandas>=1.4.1
python-telegram-bot>=12.8
